requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
discord.py-rw==1.0.0a0.post2
idna==2.8
multidict==4.5.2
Pillow==5.4.1
pychalk==2.0.1
python-dotenv==0.10.1
redis==3.1.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websockets==6.0
yarl==1.3.0
